Updated entity and new Guid id

+++ Entities/DataTransferObjects/CarForCreationDto.cs
@@ -9,10 +9,10 @@ namespace CarsCatalogue.Models.DataTransferObjects
         public string model { get; set; }
         public string year { get; set; }
         public int speed { get; set; }
-        public int accelaration { get; set; }
+        public double accelaration { get; set; }
 		public double price { get; set; }
         public string engine { get; set; }
-		public int capacity { get; set; }
+		public double capacity { get; set; }
         public string description { get; set; }
 		public string image {get; set;} 
 		public string video { get; set; }
